Added handling of different types to iOS
