+++ hooks/iosAddTarget.js
@@ -28,100 +28,39 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 //
+const fs = require('fs');
+const path = require('path');
 
-const PLUGIN_ID = 'cc.fovea.cordova.openwith';
-const BUNDLE_SUFFIX = '.shareextension';
+const {
+  PLUGIN_ID,
+  iosFolder,
+  getPreferences,
+  findXCodeproject,
+  replacePreferencesInFile,
+  log, redError,
+} = require('./utils')
 
-var fs = require('fs');
-var path = require('path');
-
-function redError(message) {
-    return new Error('"' + PLUGIN_ID + '" \x1b[1m\x1b[31m' + message + '\x1b[0m');
-}
-
-function replacePreferencesInFile(filePath, preferences) {
-    var content = fs.readFileSync(filePath, 'utf8');
-    for (var i = 0; i < preferences.length; i++) {
-        var pref = preferences[i];
-        var regexp = new RegExp(pref.key, "g");
-        content = content.replace(regexp, pref.value);
-    }
-    fs.writeFileSync(filePath, content);
-}
-
-// Determine the full path to the app's xcode project file.
-function findXCodeproject(context, callback) {
-  fs.readdir(iosFolder(context), function(err, data) {
-    var projectFolder;
-    var projectName;
-    // Find the project folder by looking for *.xcodeproj
-    if (data && data.length) {
-      data.forEach(function(folder) {
-        if (folder.match(/\.xcodeproj$/)) {
-          projectFolder = path.join(iosFolder(context), folder);
-          projectName = path.basename(folder, '.xcodeproj');
-        }
-      });
-    }
-
-    if (!projectFolder || !projectName) {
-      throw redError('Could not find an .xcodeproj folder in: ' + iosFolder(context));
-    }
-
-    if (err) {
-      throw redError(err);
-    }
-
-    callback(projectFolder, projectName);
-  });
-}
-
-// Determine the full path to the ios platform
-function iosFolder(context) {
-  return context.opts.cordova.project
-    ? context.opts.cordova.project.root
-    : path.join(context.opts.projectRoot, 'platforms/ios/');
-}
-
-function getPreferenceValue(configXml, name) {
-  var value = configXml.match(new RegExp('name="' + name + '" value="(.*?)"', "i"));
-  if (value && value[1]) {
-    return value[1];
-  } else {
-    return null;
-  }
-}
-
-function getCordovaParameter(configXml, variableName) {
-  var variable;
-  var arg = process.argv.filter(function(arg) {
-    return arg.indexOf(variableName + '=') == 0;
-  });
-  if (arg.length >= 1) {
-    variable = arg[0].split('=')[1];
-  } else {
-    variable = getPreferenceValue(configXml, variableName);
-  }
-  return variable;
-}
-
-// Get the bundle id from config.xml
-// function getBundleId(context, configXml) {
-//   var elementTree = context.requireCordovaModule('elementtree');
-//   var etree = elementTree.parse(configXml);
-//   return etree.getroot().get('id');
-// }
+// Return the list of files in the share extension project, organized by type
+const FILE_TYPES = {
+  '.h':'source',
+  '.m':'source',
+  '.plist':'config',
+  '.entitlements':'config',
+};
 
 function parsePbxProject(context, pbxProjectPath) {
   var xcode = context.requireCordovaModule('xcode');
-  console.log('    Parsing existing project at location: ' + pbxProjectPath + '...');
+  log(`Parsing existing project at location: ${pbxProjectPath}…`);
+
   var pbxProject;
+
   if (context.opts.cordova.project) {
     pbxProject = context.opts.cordova.project.parseProjectFile(context.opts.projectRoot).xcode;
   } else {
     pbxProject = xcode.project(pbxProjectPath);
     pbxProject.parseSync();
   }
+
   return pbxProject;
 }
 
@@ -131,125 +70,50 @@ function forEachShareExtensionFile(context, callback) {
     // Ignore junk files like .DS_Store
     if (!/^\..*/.test(name)) {
       callback({
-        name:name,
-        path:path.join(shareExtensionFolder, name),
-        extension:path.extname(name)
+        name: name,
+        path: path.join(shareExtensionFolder, name),
+        extension: path.extname(name)
       });
     }
   });
 }
 
-function projectPlistPath(context, projectName) {
-  return path.join(iosFolder(context), projectName, projectName + '-Info.plist');
-}
-
-function projectPlistJson(context, projectName) {
-  var plist = require('plist');
-  var path = projectPlistPath(context, projectName);
-  return plist.parse(fs.readFileSync(path, 'utf8'));
-}
-
-function getPreferences(context, configXml, projectName) {
-  var plist = projectPlistJson(context, projectName);
-  var group = "group." + plist.CFBundleIdentifier + BUNDLE_SUFFIX;
-  if (getCordovaParameter(configXml, 'GROUP_IDENTIFIER') !== "") {
-    group = getCordovaParameter(configXml, 'IOS_GROUP_IDENTIFIER');
-  }
-  return [{
-    key: '__DISPLAY_NAME__',
-    value: projectName
-  }, {
-    key: '__BUNDLE_IDENTIFIER__',
-    value: plist.CFBundleIdentifier + BUNDLE_SUFFIX
-  } ,{
-      key: '__GROUP_IDENTIFIER__',
-      value: group
-  }, {
-    key: '__BUNDLE_SHORT_VERSION_STRING__',
-    value: plist.CFBundleShortVersionString
-  }, {
-    key: '__BUNDLE_VERSION__',
-    value: plist.CFBundleVersion
-  }, {
-    key: '__URL_SCHEME__',
-    value: getCordovaParameter(configXml, 'IOS_URL_SCHEME')
-  }, {
-    key: '__UNIFORM_TYPE_IDENTIFIER__',
-    value: getCordovaParameter(configXml, 'IOS_UNIFORM_TYPE_IDENTIFIER')
-  }];
-}
-
-// Return the list of files in the share extension project, organized by type
 function getShareExtensionFiles(context) {
-  var files = {source:[],plist:[],resource:[]};
-  var FILE_TYPES = { '.h':'source', '.m':'source', '.plist':'plist' };
+  var files = { source: [], config: [], resource: [] };
+
   forEachShareExtensionFile(context, function(file) {
     var fileType = FILE_TYPES[file.extension] || 'resource';
     files[fileType].push(file);
   });
+
   return files;
 }
 
-function printShareExtensionFiles(files) {
-  console.log('    Found following files in your ShareExtension folder:');
-  console.log('    Source files:');
-  files.source.forEach(function(file) {
-    console.log('     - ', file.name);
-  });
-
-  console.log('    Plist files:');
-  files.plist.forEach(function(file) {
-    console.log('     - ', file.name);
-  });
-
-  console.log('    Resource files:');
-  files.resource.forEach(function(file) {
-    console.log('     - ', file.name);
-  });
-}
-
-console.log('Adding target "' + PLUGIN_ID + '/ShareExtension" to XCode project');
-
-module.exports = function (context) {
+module.exports = function(context) {
+  log('Adding ShareExt target to XCode project')
 
   var Q = context.requireCordovaModule('q');
   var deferral = new Q.defer();
 
-  // if (context.opts.cordova.platforms.indexOf('ios') < 0) {
-  //   log('You have to add the ios platform before adding this plugin!', 'error');
-  // }
-
-  var configXml = fs.readFileSync(path.join(context.opts.projectRoot, 'config.xml'), 'utf-8');
-  if (configXml) {
-    configXml = configXml.substring(configXml.indexOf('<'));
-  }
-
   findXCodeproject(context, function(projectFolder, projectName) {
-
-    console.log('  - Folder containing your iOS project: ' + iosFolder(context));
+    var preferences = getPreferences(context, projectName);
 
     var pbxProjectPath = path.join(projectFolder, 'project.pbxproj');
     var pbxProject = parsePbxProject(context, pbxProjectPath);
 
     var files = getShareExtensionFiles(context);
-    // printShareExtensionFiles(files);
-
-    var preferences = getPreferences(context, configXml, projectName);
-    files.plist.concat(files.source).forEach(function(file) {
+    files.config.concat(files.source).forEach(function(file) {
       replacePreferencesInFile(file.path, preferences);
-      // console.log('    Successfully updated ' + file.name);
     });
 
     // Find if the project already contains the target and group
     var target = pbxProject.pbxTargetByName('ShareExt');
-    if (target) {
-      console.log('    ShareExt target already exists.');
-    }
+    if (target) { log('ShareExt target already exists') }
 
     if (!target) {
       // Add PBXNativeTarget to the project
       target = pbxProject.addTarget('ShareExt', 'app_extension', 'ShareExtension');
-      
+
       // Add a new PBXSourcesBuildPhase for our ShareViewController
       // (we can't add it to the existing one because an extension is kind of an extra app)
       pbxProject.addBuildPhase([], 'PBXSourcesBuildPhase', 'Sources', target.uuid);
@@ -261,9 +125,8 @@ module.exports = function (context) {
 
     // Create a separate PBXGroup for the shareExtensions files, name has to be unique and path must be in quotation marks
     var pbxGroupKey = pbxProject.findPBXGroupKey({name: 'ShareExtension'});
-    if (pbxProject) {
-      console.log('    ShareExtension group already exists.');
-    }
+    if (pbxProject) { log('ShareExtension group already exists') }
+
     if (!pbxGroupKey) {
       pbxGroupKey = pbxProject.pbxCreateGroup('ShareExtension', 'ShareExtension');
 
@@ -273,7 +136,7 @@ module.exports = function (context) {
     }
 
     // Add files which are not part of any build phase (config)
-    files.plist.forEach(function (file) {
+    files.config.forEach(function (file) {
       pbxProject.addFile(file.name, pbxGroupKey);
     });
 
@@ -287,77 +150,23 @@ module.exports = function (context) {
       pbxProject.addResourceFile(file.name, {target: target.uuid}, pbxGroupKey);
     });
 
-    //Add development team and provisioning profile
-    var PROVISIONING_PROFILE = getCordovaParameter(configXml, 'SHAREEXT_PROVISIONING_PROFILE');
-    var DEVELOPMENT_TEAM = getCordovaParameter(configXml, 'SHAREEXT_DEVELOPMENT_TEAM');
-    console.log('Adding team', DEVELOPMENT_TEAM, 'and provisoning profile', PROVISIONING_PROFILE);
-    if (PROVISIONING_PROFILE && DEVELOPMENT_TEAM) {
-      var configurations = pbxProject.pbxXCBuildConfigurationSection();
-      for (var key in configurations) {
-        if (typeof configurations[key].buildSettings !== 'undefined') {
-          var buildSettingsObj = configurations[key].buildSettings;
-          if (typeof buildSettingsObj['PRODUCT_NAME'] !== 'undefined') {
-            var productName = buildSettingsObj['PRODUCT_NAME'];
-            if (productName.indexOf('ShareExt') >= 0) {
-              buildSettingsObj['PROVISIONING_PROFILE'] = PROVISIONING_PROFILE;
-              buildSettingsObj['DEVELOPMENT_TEAM'] = DEVELOPMENT_TEAM;
-              console.log('Added signing identities for extension!');
-            }
+    // Add build settings for Swift support, bridging header and xcconfig files
+    var configurations = pbxProject.pbxXCBuildConfigurationSection();
+    for (var key in configurations) {
+      if (typeof configurations[key].buildSettings !== 'undefined') {
+        var buildSettingsObj = configurations[key].buildSettings;
+        if (typeof buildSettingsObj['PRODUCT_NAME'] !== 'undefined') {
+          var productName = buildSettingsObj['PRODUCT_NAME'];
+          if (productName.indexOf('ShareExt') >= 0) {
+            buildSettingsObj['CODE_SIGN_ENTITLEMENTS'] = '"ShareExtension/ShareExtension.entitlements"';
           }
         }
       }
     }
 
-    // Add a new PBXFrameworksBuildPhase for the Frameworks used by the Share Extension
-    // (NotificationCenter.framework, libCordova.a)
-    // var frameworksBuildPhase = pbxProject.addBuildPhase(
-    //   [],
-    //   'PBXFrameworksBuildPhase',
-    //   'Frameworks',
-    //   target.uuid
-    // );
-    // if (frameworksBuildPhase) {
-    //   log('Successfully added PBXFrameworksBuildPhase!', 'info');
-    // }
-
-    // Add the frameworks needed by our shareExtension, add them to the existing Frameworks PbxGroup and PBXFrameworksBuildPhase
-    // var frameworkFile1 = pbxProject.addFramework(
-    //   'NotificationCenter.framework',
-    //   { target: target.uuid }
-    // );
-    // var frameworkFile2 = pbxProject.addFramework('libCordova.a', {
-    //   target: target.uuid,
-    // }); // seems to work because the first target is built before the second one
-    // if (frameworkFile1 && frameworkFile2) {
-    //   log('Successfully added frameworks needed by the share extension!', 'info');
-    // }
-
-    // Add build settings for Swift support, bridging header and xcconfig files
-    // var configurations = pbxProject.pbxXCBuildConfigurationSection();
-    // for (var key in configurations) {
-    //   if (typeof configurations[key].buildSettings !== 'undefined') {
-    //     var buildSettingsObj = configurations[key].buildSettings;
-    //     if (typeof buildSettingsObj['PRODUCT_NAME'] !== 'undefined') {
-    //       var productName = buildSettingsObj['PRODUCT_NAME'];
-    //       if (productName.indexOf('ShareExtension') >= 0) {
-    //         if (addXcconfig) {
-    //           configurations[key].baseConfigurationReference =
-    //             xcconfigReference + ' /* ' + xcconfigFileName + ' */';
-    //           log('Added xcconfig file reference to build settings!', 'info');
-    //         }
-    //         if (addEntitlementsFile) {
-    //           buildSettingsObj['CODE_SIGN_ENTITLEMENTS'] = '"' + 'ShareExtension' + '/' + entitlementsFileName + '"';
-    //           log('Added entitlements file reference to build settings!', 'info');
-    //         }
-    //       }
-    //     }
-    //   }
-    // }
-
     // Write the modified project back to disc
-    // console.log('    Writing the modified project back to disk...');
     fs.writeFileSync(pbxProjectPath, pbxProject.writeSync());
-    console.log('Added ShareExtension to XCode project');
+    log('Successfully added ShareExt target to XCode project')
 
     deferral.resolve();
   });
