+++ hooks/iosCopyShareExtension.js
@@ -1,55 +1,15 @@
-//
-//  iosCopyShareExtension.js
-//  This hook runs for the iOS platform when the plugin or platform is added.
-//
-// Source: https://github.com/DavidStrausz/cordova-plugin-today-widget
-//
+const fs = require('fs');
+const path = require('path');
 
-//
-// The MIT License (MIT)
-//
-// Copyright (c) 2017 DavidStrausz
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-//
+const {
+  PLUGIN_ID,
+  getPreferences,
+  findXCodeproject,
+  replacePreferencesInFile,
+  log, redError,
+} = require('./utils')
 
-var fs = require('fs');
-var path = require('path');
-const PLUGIN_ID = "cc.fovea.cordova.openwith";
-
-function redError(message) {
-    return new Error('"' + PLUGIN_ID + '" \x1b[1m\x1b[31m' + message + '\x1b[0m');
-}
-
-function getPreferenceValue (config, name) {
-  var value = config.match(new RegExp('name="' + name + '" value="(.*?)"', "i"));
-  if(value && value[1]) {
-    return value[1];
-  } else {
-    return null;
-  }
-}
-
-console.log('Copying "' + PLUGIN_ID + '/ShareExtension" to ios...');
-
-// http://stackoverflow.com/a/26038979/5930772
-function copyFileSync(source, target) {
+function copyFileSync(source, target, preferences) {
   var targetFile = target;
 
   // If target is a directory a new file with the same name will be created
@@ -60,9 +20,10 @@ function copyFileSync(source, target) {
   }
 
   fs.writeFileSync(targetFile, fs.readFileSync(source));
+  replacePreferencesInFile(targetFile, preferences);
 }
 
-function copyFolderRecursiveSync(source, target) {
+function copyFolderRecursiveSync(source, target, preferences) {
   var files = [];
 
   // Check if folder needs to be created or integrated
@@ -77,56 +38,31 @@ function copyFolderRecursiveSync(source, target) {
     files.forEach(function(file) {
       var curSource = path.join(source, file);
       if (fs.lstatSync(curSource).isDirectory()) {
-        copyFolderRecursiveSync(curSource, targetFolder);
+        copyFolderRecursiveSync(curSource, targetFolder, preferences);
       } else {
-        copyFileSync(curSource, targetFolder);
+        copyFileSync(curSource, targetFolder, preferences);
       }
     });
   }
 }
 
-// Determine the full path to the app's xcode project file.
-function findXCodeproject(context, callback) {
-  var iosFolder = context.opts.cordova.project
-    ? context.opts.cordova.project.root
-    : path.join(context.opts.projectRoot, 'platforms/ios/');
-  fs.readdir(iosFolder, function(err, data) {
-    var projectFolder;
-    var projectName;
-    // Find the project folder by looking for *.xcodeproj
-    if (data && data.length) {
-      data.forEach(function(folder) {
-        if (folder.match(/\.xcodeproj$/)) {
-          projectFolder = path.join(iosFolder, folder);
-          projectName = path.basename(folder, '.xcodeproj');
-        }
-      });
-    }
-
-    if (!projectFolder || !projectName) {
-      throw redError('Could not find an .xcodeproj folder in: ' + iosFolder);
-    }
-
-    if (err) {
-      throw redError(err);
-    }
-
-    callback(projectFolder, projectName);
-  });
-}
-
 module.exports = function(context) {
+  log('Copying ShareExtension files to iOS project')
+
   var Q = context.requireCordovaModule('q');
   var deferral = new Q.defer();
 
   findXCodeproject(context, function(projectFolder, projectName) {
+    var preferences = getPreferences(context, projectName);
 
     var srcFolder = path.join(context.opts.projectRoot, 'plugins', PLUGIN_ID, 'src', 'ios', 'ShareExtension');
+    var targetFolder = path.join(context.opts.projectRoot, 'platforms', 'ios');
+
     if (!fs.existsSync(srcFolder)) {
       throw redError('Missing extension project folder in ' + srcFolder + '.');
     }
 
-    copyFolderRecursiveSync(srcFolder, path.join(context.opts.projectRoot, 'platforms', 'ios'));
+    copyFolderRecursiveSync(srcFolder, targetFolder, preferences);
     deferral.resolve();
   });
