+++ src/ios/ShareExtension/ShareExtension-Info.plist
@@ -20,29 +20,5 @@
 	<string>__BUNDLE_SHORT_VERSION_STRING__</string>
 	<key>CFBundleVersion</key>
 	<string>__BUNDLE_VERSION__</string>
-	<key>NSExtension</key>
-	<dict>
-		<key>NSExtensionAttributes</key>
-		<dict>
-			<key>NSExtensionActivationRule</key>
-			<!-- string>TRUEPREDICATE</string -->
-      <dict>
-        <key>NSExtensionActivationSupportsImageWithMaxCount</key>
-        <integer>1</integer>
-        <!-- key>NSExtensionActivationSupportsMovieWithMaxCount</key>
-        <integer>1</integer -->
-        <!-- key>NSExtensionActivationSupportsWebURLWithMaxCount</key>
-        <integer>1</integer -->
-      </dict>
-      <key>NSExtensionActivationUsesStrictMatching</key>
-      <integer>1</integer>
-      <key>NSExtensionActivationDictionaryVersion</key>
-      <integer>2</integer>
-		</dict>
-		<key>NSExtensionMainStoryboard</key>
-		<string>MainInterface</string>
-		<key>NSExtensionPointIdentifier</key>
-		<string>com.apple.share-services</string>
-	</dict>
 </dict>
 </plist>
