+++ src/ios/ShareExtension/ShareViewController.h
@@ -25,6 +25,5 @@
 // THE SOFTWARE.
 //
 
-#define SHAREEXT_GROUP_IDENTIFIER @"__GROUP_IDENTIFIER__"
+#define SHAREEXT_GROUP_IDENTIFIER @"group.__BUNDLE_IDENTIFIER__"
 #define SHAREEXT_URL_SCHEME @"__URL_SCHEME__"
-#define SHAREEXT_UNIFORM_TYPE_IDENTIFIER @"__UNIFORM_TYPE_IDENTIFIER__"
